demos/RobotFramework/requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
bcrypt==3.1.6
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
idna==2.8
ipaddress==1.0.22
jdcal==1.4.1
jsonpatch==1.23
jsonpointer==2.0
natsort==6.0.0
ndg-httpsclient==0.5.1
openpyxl==2.6.2
paramiko==2.4.2
Pillow==6.0.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyYAML==5.1
requests==2.22.0
robotframework==3.1.2
robotframework-angularjs==0.0.9
robotframework-excellibrary==0.0.2
robotframework-httplibrary==0.4.2
robotframework-requests==0.5.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1
robotframework-sshlibrary==3.3.0
robotframework-testrail==1.0.4
robotframework-xvfb==1.2.2
scp==0.13.2
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
waitress==1.3.0
WebOb==1.8.5
WebTest==2.0.33
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
xvfbwrapper==0.2.9
dotdict==0.1
ordereddict==1.1
WooCommerce==2.1.1
